+++ client/src/hooks/index.js
@@ -1,0 +1,1 @@
+export { usePosts } from './usePosts';
